add --use-right-blanks, use _ for missing, fix relative error when mean 0
The --use-right-blanks option computes the statistics for the left side
variables by omitting the periods of missing values in the right side, so
that bad periods in the left side are not used to compare against periods
in the right which have since been blanked out.  Print _ for missing values
when printing variable values in output with indexes.  If the absolute
error is zero, then set relative error to zero.  Now relative error is not
set to 100% when both means are zero and thus should compare as equal.