Remove --prefix
Moving away from that construct for the time being.

In eol_scons/tools/prefixopttions.py DEFAULT_INSTALL_PREFIX defaults to /usr/local for Mac and /opt/local for all else.  No need at this time for installion in place.

If we want to add back to --prefix, it should be done once in prefixoptions.py, not replicated in top level SConstructs.