Refactored HRT and SRT section after discussion with Chris and Janine.
Main change: store rate in a dictionary and determine alignment based on
max rate in subselected variables. I was pretty repetitive (DRY!. I
am still keeping the flag -mixed_rate, because if the file is low rate,
a lot of other bells and whistles get executed. The code is rate agnostic though,
so if you use this flag on a LRT file, it will work.