fix and feature related to non-normal floats
add the --nans-equal flag to allow two nans to compare as equal, meaning a nan in the right file
is not a difference if there is a nan in the left file.  also discovered a bug where inf numbers would
compare as equal with ulps but not with delta, so now they compare as equal for both.

for context, profiler netcdf files output from nima can contain lots of nans and infs.