handle variables of different types w/out crashing
If the right variable cannot be cast to the same nc_var<T> as the left, then
use the new virtual getDouble() method to compare the values as doubles.
Perhaps not the most efficient way to compare different types, but effective,
and it provides more information than just skipping the variable.