replace gsl with simple template implementation
since gsl has statistics functions which take type long, but type long may not
be the same as int64 on some platforms (mac arm?), it seems simpler to replace
gsl calls with a single template which can be instantiated exactly with the
needed type.

while we're at it, migrate from boost to std:: smart pointers, and add
arguments to the SConscript to specify different test builds.